refactor(tui): make TUI mutation helpers update-only.
